feat(explorer): add basic proposal vote page
